Fixed bug in room tile traversal.
Was choosing the start/end tile sides for the
inner axis incorrectly. Changed it to the reverse.

Now the function never appears to get too /few/
tiles, but there are cases where it returns too
many due to coosing between the calculated and
ultimate start/end points with the slope.