Fixed a bug in room's tile traversal.
I believe the function should be 100%
working now. It appears to be passing
all the test cases I've tried.

Now that getting inner and outer
directions had been added, it was
only needed to swap the slope > 0.0f
check to a isMax(innerDir) check to
determine whether to use the
calculated position or the min/max
position.