Added Polygon to Polygon intersection test.
Can test Polygon to Polygon, and thus any
Shape by first converting to a polygon.

This intersection function uses SAT, and will
return false if the polygons "touch" but have
no overlap. This might seem inconsistent with
my other intersection functions, so I might
need to consider changing that either in this
one or the others.

I think for actual collisions in games, ignoring
when they just touch is probably better. The only
edge case that comes to mind for this is if an
object is trying to squeeze between two other
objects.