Added preliminary collision tests.
Works most of the time, but has edge cases that are easy to encounter
that break the system.

Basically if the result of the collision resolution doesn't push the
collider completely out of the polygon, or pushes the collider into
another nearby polygon, then the next update may have odd behaviour.

Generally easy to see by wedging between two polygons and jittering
back and forth. Eventually the collider will be pushed inside one
of them, and then going a different direction may have it "warp" to
the other side.