Fixed an edge case with collision math.
Determined how to deal with an edge case where the
SAT test claims that the polygons are intersecting,
but the penetration test finds no overlap.

Just assume it hits something and stop it,
essentially, pushing it back a bit to ensure that
the SAT test should no longer be positive.