Fixed oscillation problem when entering wedges.
Could end up with very deep recursions when moving
into a wedge shape. Was handling this before simply
by cutting off recursions after a point.

Now keeps track of the minimum and maximum angle between
the deflection directions and original delta direction.
If the current deflection angle is between the prevoius
min and max, and the polygon did not move during this
recursion level, then we can assume we are stuck and
that further recusions are not needed.