Sped up line segment and ray intersection tests.
Added some bounds tests for segment-segment and
ray-segment intersection tests.

When tested against a large number of intersections,
the segment-segment test was barely changed at all,
while the ray-segment test shows significant
improvement.

Also verified that the no-output-point version of the
segment-segment test is indeed much faster than the
version that finds an output point, and they appear
to have the same accuraccy.