Bugfix: hybridSAT finds future overlap.
Fixed bug where hybrid SAT could falsely return
there being a collision when polygons are merely
touching, not overlapping.

This is an extremely minor bug affecting the edge
case where a polygon moves to be exactly touching
another polygon -- but not overlapping it. In such
a case we wish to say there is no collision.